Fixed UI datasim implementation

+++ app/src/main/java/com/example/wpms/ViewModel/PatientHomeActivityViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.wpms.ViewModel
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -10,7 +9,7 @@ import java.sql.Timestamp
 class PatientHomeActivityViewModel(private val firebaseRepository: FirebaseRepository): ViewModel() {
     private val pressureData = MutableLiveData<List<Int>>()
 //    val pressureData: LiveData<List<Int>> get() = pressureData
-    fun insertPressureData(deviceID: String, pressure_center: Int, pressure_left: Int, pressure_right: Int, timestamp: Timestamp) {
+    fun insertPressureData(deviceID: String, pressure_center: Float, pressure_left: Float, pressure_right: Float, timestamp: Timestamp) {
         // Insert pressure data into Firestore
         firebaseRepository.insertPressureData(deviceID, pressure_center, pressure_left, pressure_right, timestamp)
     }
